+++ HabboHotel/Items/Wired/Boxes/Conditions/FurniMatchStateAndPositionBox.cs
@@ -62,16 +62,16 @@ internal class FurniMatchStateAndPositionBox : IWiredItem
             {
                 if (string.IsNullOrEmpty(I))
                     continue;
-                var ii = Instance.GetRoomItemHandler().GetItem(Convert.ToUInt32(I.Split(':')[0]));
+                if (!WiredFurniSnapshotParser.TryParseEntry(I, out var itemId, out var snapshot))
+                    continue;
+                var ii = Instance.GetRoomItemHandler().GetItem(itemId);
                 if (ii == null)
                     continue;
-                var partsString = I.Split(':');
-                var part = partsString[1].Split(',');
                 if (stateMode == 1) //State
                 {
                     try
                     {
-                        if (ii.LegacyDataString != part[4])
+                        if (ii.LegacyDataString != snapshot.State)
                             return false;
                     }
                     catch { }
@@ -80,7 +80,7 @@ internal class FurniMatchStateAndPositionBox : IWiredItem
                 {
                     try
                     {
-                        if (ii.Rotation != Convert.ToInt32(part[3]))
+                        if (ii.Rotation != snapshot.Rotation)
                             return false;
                     }
                     catch { }
@@ -89,8 +89,7 @@ internal class FurniMatchStateAndPositionBox : IWiredItem
                 {
                     try
                     {
-                        if (ii.GetX != Convert.ToInt32(part[0]) || ii.GetY != Convert.ToInt32(part[1]) ||
-                            ii.GetZ != Convert.ToDouble(part[2]))
+                        if (ii.GetX != snapshot.X || ii.GetY != snapshot.Y || ii.GetZ != snapshot.Z)
                             return false;
                     }
                     catch { }